LayoutTests/fast/events/pageshow-pagehide-on-back-cached-expected.txt

ALERT: window.onload
ALERT: window.onpageshow, target = [object HTMLDocument], persisted = false
ALERT: window.onpagehide, target = [object HTMLDocument], persisted = true
ALERT: window.onpageshow, target = [object HTMLDocument], persisted = true
Test pageshow/pagehide event behavior when navigating back to an cached page.
